+++ src/GlobalPayments.Api/PaymentMethods/PaymentInterfaces/ICardData.cs
@@ -11,5 +11,6 @@ namespace GlobalPayments.Api.PaymentMethods {
         int? ExpYear { get; set; }
         bool ReaderPresent { get; set; }
         string ShortExpiry { get; }
+        ManualEntryMethod? EntryMethod { get; set; }
     }
 }